Images added and Example code fixed to resolve errors
- Added the image folders from one drive but they still need to be transformed (greyscale and cropped via the transform_images module) consider adding that function to the main/example module
- Example code fixed enough to resolve errors and output a value
- Example.py still needs to be reworked from the original 1 image then set of images to the 30 / 70 sets of images
- also need to investigate if the PRNU_Ref_Image.pdf is still solid black after reworking the file

+++ PRNU_MATH535_PYTHON/src/Filter.py
@@ -158,9 +158,9 @@ def NoiseExtractFromImage(image, sigma=2.0, color=False, noZM=False):
     M0, N0  = X.shape
     if X.dtype == 'uint8':
         # convert to [0,255]
-        X = X.astype(np.float)
+        X = X.astype(np.float32)
     elif X.dtype == 'uint16':
-        X = X.astype(np.float) / 65535 * 255
+        X = X.astype(np.float32) / 65535 * 255
 
     qmf = [ 	.230377813309,	.714846570553, .630880767930, -.027983769417,
            -.187034811719,	.030841381836, .032883011667, -.010597401785]
